move AppContext to  hooks
