+++ src/components/ProgressDisplay.tsx
@@ -1,4 +1,4 @@
-import { AppContext } from "../models";
+import { AppContext } from "../hooks";
 
 export default function ProgressDisplay(context: AppContext) {
     const { elementsHook, stateHook } = context;
