+++ src/events/onResult.tsx
@@ -1,6 +1,5 @@
-import { GameState, ButtonState } from "../enums";
-import { AppContext } from "../models";
-
+import { AppContext } from "../hooks";
+import { ButtonState, GameState } from "../enums";
 ///
 var wrongGuesses: number[] = [];
