+++ src/hooks/useEventRouter.tsx
@@ -1,7 +1,7 @@
-import { AppContext } from "../models";
 import { GameState } from "../enums";
 import { onInit, onLoading, onNext, onInput, onResult, onGameOver } from "../events";
 import { useEffect } from "react";
+import AppContext from "./AppContext";
 
 type EventHandlers = {
     [key in GameState]: (context: AppContext) => void;
