+++ src/models/index.tsx
@@ -1,8 +1,7 @@
-import AppContext from "./AppContext";
 import AppState from "./AppState";
 import Config from "./Config";
 import { ConfigDefaults } from "./ConfigDefaults";
 import { QuizModule, QuizData, QuizItem } from "./QuizModule";
 
 export { ConfigDefaults };
-export type { AppContext, AppState, Config, QuizModule, QuizData, QuizItem };
+export type { AppState, Config, QuizModule, QuizData, QuizItem };
